Enable -Werror and fix all warnings
Build system:
- Add -Werror to all library targets (sparkle-duck-core, sparkle-duck-server-lib, sparkle-duck-ui-lib)
- Build now fails on any warnings, catching issues early

Code cleanup:
- Remove unused com_mass_constant property from MaterialType
- Remove unused variables in Tree.cpp (center_x, center_y)
- Mark intentionally unused parameters with [[maybe_unused]]
- Fix sign comparison warnings (int vs uint32_t)
- Fix narrowing conversion warning (double to uint32_t)

API improvements:
- Add name() method to SimRun::Command for consistency
- Add sendCommand() template method to WebSocketClient
- Preserve max_frame_ms when switching scenarios

Test updates:
- Update Buoyancy tests to use sandbox default physics settings

+++ test-lvgl/src/ui/rendering/CellRenderer.cpp
@@ -280,8 +280,8 @@ void CellRenderer::renderWorldData(
             "CellRenderer: Total world rendering size {}x{}, fits in canvas? W:{} H:{}",
             totalW,
             totalH,
-            totalW <= canvasWidth_,
-            totalH <= canvasHeight_);
+            static_cast<uint32_t>(totalW) <= canvasWidth_,
+            static_cast<uint32_t>(totalH) <= canvasHeight_);
         lastWidth = worldData.width;
         lastHeight = worldData.height;
     }
